ciscolive-demo/documentation-llm/requirements.txt
accelerate==0.24.1
aioconsole==0.6.2
aiohttp==3.8.5
aiosignal==1.3.1
aiostream==0.5.2
annotated-types==0.5.0
antlr4-python3-runtime==4.9.3
anyio==3.7.1
appdirs==1.4.4
argilla==1.18.0
astunparse==1.6.3
async-timeout==4.0.3
attrs==23.1.0
backoff==2.2.1
beautifulsoup4==4.12.2
bert-score==0.3.13
bidict==0.22.1
bitsandbytes==0.41.2.post2
blinker==1.6.2
cachetools==5.3.2
certifi==2023.7.22
cffi==1.15.1
chardet==5.2.0
charset-normalizer==3.2.0
click==8.1.7
coloredlogs==15.0.1
contourpy==1.1.1
cryptography==41.0.3
ctransformers==0.2.25
cycler==0.12.0
dataclasses-json==0.5.14
datasets==2.14.6
Deprecated==1.2.14
dill==0.3.7
diskcache==5.6.3
distro==1.8.0
dnspython==2.4.2
docker-pycreds==0.4.0
effdet==0.4.1
emoji==2.8.0
et-xmlfile==1.1.0
evaluate==0.4.1
eventlet==0.33.3
exceptiongroup==1.1.3
fastapi==0.104.1
filelock==3.12.2
filetype==1.2.0
Flask==2.3.3
Flask-Cors==4.0.0
Flask-SocketIO==5.3.6
flatbuffers==23.5.26
fonttools==4.43.0
frozenlist==1.4.0
fsspec==2023.6.0
gitdb==4.0.10
GitPython==3.1.32
gptcache==0.1.42
greenlet==2.0.2
guidance==0.1.2
h11==0.14.0
hiredis==2.2.3
httpcore==0.18.0
httpx==0.25.0
huggingface-hub==0.16.4
humanfriendly==10.0
idna==3.4
iniconfig==2.0.0
iopath==0.1.10
itsdangerous==2.1.2
Jinja2==3.1.2
joblib==1.3.2
jsonpatch==1.33
jsonpointer==2.4
kiwisolver==1.4.5
langchain==0.0.309
langchainhub==0.1.13
langdetect==1.0.9
langsmith==0.0.42
layoutparser==0.3.4
llama-cpp-guidance==0.1.2
llama-index==0.8.62
llama_cpp_python==0.2.19
lmql==0.7.3
loguru==0.7.2
lxml==4.9.3
Markdown==3.5.1
markdown-it-py==3.0.0
MarkupSafe==2.1.3
marshmallow==3.20.1
matplotlib==3.8.0
mdurl==0.1.2
monotonic==1.6
more-itertools==9.1.0
mpmath==1.3.0
msal==1.24.1
msg-parser==1.2.0
multidict==6.0.4
multiprocess==0.70.15
mypy-extensions==1.0.0
nest-asyncio==1.5.8
networkx==3.1
nltk==3.8.1
numpy==1.26.2
olefile==0.46
omegaconf==2.3.0
onnx==1.15.0
onnxruntime==1.15.1
openai==1.3.4
opencv-python==4.8.1.78
openpyxl==3.1.2
ordered-set==4.1.0
outcome==1.2.0
packaging==23.1
pandas==2.1.1
pathtools==0.1.2
pdf2image==1.16.3
pdfminer.six==20221105
pdfplumber==0.10.2
peft==0.6.2
Pillow==10.0.0
platformdirs==3.11.0
pluggy==1.2.0
portalocker==2.8.2
protobuf==4.24.1
psutil==5.9.5
psycopg2-binary==2.9.7
py-cpuinfo==9.0.0
pyarrow==14.0.0
pycocotools==2.0.7
pycparser==2.21
pycryptodome==3.19.0
pydantic==2.4.2
pydantic-settings==2.0.3
pydantic_core==2.10.1
pydot==1.4.2
pyformlang==1.0.4
Pygments==2.16.1
pygtrie==2.5.0
PyJWT==2.8.0
pypandoc==1.12
pyparsing==3.1.1
PyPDF2==3.0.1
pypdfium2==4.20.0
pysbd==0.3.4
PySocks==1.7.1
pytesseract==0.3.10
pytest==7.4.0
pytest-asyncio==0.21.1
python-dateutil==2.8.2
python-docx==1.0.1
python-dotenv==1.0.0
python-engineio==4.7.1
python-iso639==2023.6.15
python-magic==0.4.27
python-multipart==0.0.6
python-pptx==0.6.21
python-socketio==5.9.0
python-ulid==1.1.0
pytz==2023.3.post1
PyYAML==6.0.1
ragas==0.0.19
rapidfuzz==3.5.1
redis==4.6.0
redis-om==0.2.1
regex==2023.8.8
requests==2.31.0
responses==0.18.0
rfc3986==1.5.0
rich==13.6.0
rwkv==0.8.20
safetensors==0.3.3
scikit-learn==1.3.0
scipy==1.11.2
selenium==4.11.2
sentence-transformers==2.2.2
sentencepiece==0.1.99
sentry-sdk==1.29.2
setproctitle==1.3.2
simple-websocket==0.10.1
six==1.16.0
sklearn==0.0.post9
smmap==5.0.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
SQLAlchemy==2.0.21
sse-starlette==1.6.5
starlette==0.27.0
starlette-context==0.3.6
sympy==1.12
tabulate==0.9.0
tenacity==8.2.3
termcolor==2.3.0
threadpoolctl==3.2.0
tiktoken==0.5.1
timm==0.9.8
tokenizers==0.15.0
tomli==2.0.1
torch==2.0.1
torchvision==0.15.2
tqdm==4.64.1
transformers==4.35.2
trio==0.22.2
trio-websocket==0.10.3
typer==0.7.0
types-pyOpenSSL==23.3.0.0
types-redis==4.6.0.8
types-requests==2.31.0.10
typing-inspect==0.9.0
typing_extensions==4.8.0
tzdata==2023.3
unstructured==0.5.6
unstructured-inference==0.7.10
unstructured.pytesseract==0.3.12
urllib3==2.0.4
uvicorn==0.23.2
wandb==0.15.5
Werkzeug==2.3.7
wrapt==1.15.0
wsproto==1.2.0
xlrd==2.0.1
XlsxWriter==3.1.9
xxhash==3.4.1
yarl==1.9.2
